requirements.txt
diff-match-patch==20121119
Django==1.9.7
django-debug-toolbar==1.4
django-import-export==0.4.5
fake-factory==0.5.9
html2text==2016.5.29
pipdeptree==0.6.0
python-dateutil==2.5.3
six==1.10.0
sqlparse==0.1.19
tablib==0.11.2
trafaret==0.7.1
